parsging ambient light, created get_uint8()

+++ incs/colours.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   colours.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hmontoya <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 18:25:40 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/11/18 18:39:48 by hmontoya         ###   ########.fr       */
+/*   Updated: 2024/11/21 20:45:05 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,19 +17,19 @@
 
 typedef struct s_color
 {
-	uint8_t r;
-	uint8_t g;
-	uint8_t b;
-	uint32_t (*normalize)(uint8_t, uint8_t, uint8_t);
+	unsigned int r;
+	unsigned int g;
+	unsigned int b;
+	uint32_t (*normalize)(unsigned int, unsigned int, unsigned int);
 }	t_color;
 
 typedef struct s_acolor
 {
-	uint8_t r;
-	uint8_t g;
-	uint8_t b;
-	uint8_t a;
-	uint32_t (*normalize)(uint8_t, uint8_t, uint8_t, uint8_t);
+	unsigned int r;
+	unsigned int g;
+	unsigned int b;
+	unsigned int a;
+	uint32_t (*normalize)(unsigned int, unsigned int, unsigned int, unsigned int);
 }	t_acolor;
